requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.8
django-environ==0.8.1
django-filter==21.1
django-storages==1.12.2
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
